The following changes were implemented:
- **Implemented Ownerless Stadium Support**: Introduced a new mode for stadiums created without an assigned owner. These stadiums now default to an "offline" booking mode, where the UI replaces the scheduling/booking flow with a direct "Call to Book" action.
- **Enhanced Stadium Management**:
    - Updated `AddPitch` and `EditStadium` screens to allow Super Admins and District Admins to manually assign an owner from a searchable dropdown.
    - Added a `phone` field to stadiums to facilitate direct contact.
    - Updated `StadiumDetailsScreen` to conditionally display booking slots or a phone contact CTA based on the presence of an owner.
- **Platform-Specific Enhancements**: Added a cross-platform `makePhoneCall` utility to `core/platform`, implementing the native dialer intent for both Android and iOS.
- **API & Data Model Updates**:
    - Updated `StadiumRepository` and UseCases to support optional `ownerId` and new `phone` parameters during creation and updates.
    - Modified `CreateStadiumRequest` and `StadiumResponse` DTOs to include the new phone and owner fields.
    - Updated `TournamentRequestDto` to include `districtId` and migrated `entryFee` from `Double` to `Long`.
- **UI & Navigation Refactoring**:
    - Restricted stadium editing capabilities in `StadiumDetailsScreen` based on user roles (`Admin` or `Owner`).
    - Updated `HomeScreen` quick actions and `MainScreen` tab navigation to streamline the user experience, including temporary removal of coach-related UI elements.
    - Integrated `UserRepository` into stadium ViewModels to fetch available owners for admin assignment.
- **Bug Fixes & Improvements**: Standardized date and time parsing in `EditStadiumViewModel` to prevent crashes on empty or malformed ISO strings.

+++ composeApp/src/commonMain/kotlin/uz/coder/foottopbusiness/presentation/main/tournaments/TournamentsViewModel.kt
@@ -44,7 +44,7 @@ class TournamentsViewModel(
                                 startDate = event.startDate,
                                 endDate = event.endDate,
                                 maxTeams = event.maxTeams,
-                                entryFee = event.entryFee,
+                                entryFee = event.entryFee.toLong(),
                                 address = event.address,
                                 startTime = event.startTime,
                                 endTime = event.endTime,
